Set serial fd to blocking in send_kernel
In the main loop, we will keep listening to stdin and serial port in the
meantime, so we need to set serial port's fd to non-blocking.

But when we need to send the kernel image to raspi, calling write on a
non-blocking fd may result in `EAGAIN` to occur.

To solve this problem, we set the serial fd to blocking in send_kernel
function and restore to non-blocking one when we leave that function. So
we will not raise any exception when sending kernel data.